application/Docker/Zone_Frost/requirements.txt
google-cloud-bigquery==3.30.0
google-cloud-bigquery-storage==2.27.0
google-cloud-storage==2.19.0
pandas==2.2.3
numpy==2.1.3
pyarrow==18.1.0
db-dtypes
torch==2.5.1
